move calculators out into their own module
